requirements.txt
# Runtime
python-dotenv>=1.0.1
PyYAML>=6.0.2
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.24.0
apscheduler>=3.10.1
oracledb>=1.0.0
